requirements.txt
django-cms>=3.6,<3.7
djangocms-admin-style>=1.2,<1.3
django-treebeard>=4.0,<5.0

djangocms-text-ckeditor>=3.7,<3.8
djangocms-link>=2.3
djangocms-style>=2.1
djangocms-googlemap>=1.2
djangocms-snippet>=2.1,<2.2
djangocms-video>=2.0,<2.1
djangocms-column>=1.9
djangocms-file>=2.2,<3.0
djangocms-picture>=2.0,<2.1
easy_thumbnails
django-filer>=1.3
Django<2.0
django-mptt>0.9
django-classy-tags>=0.7
html5lib>=1.0.1
Pillow>=3.0
django-sekizai>=0.9
six
pytz
django-crispy-forms==1.7.2
django-autoslug-iplweb==1.9.4
django-phonenumber-field
phonenumbers
aldryn-background-image==0.1.1
djangocms-background-media==1.0.0
cmsplugin-iframe


